🎨 refactor: Modularize and structure backend code with APIRouter
- Separate the backend code into modular components for better maintainability and scalability
- Move all API endpoints to api.py and utilized APIRouter for routing
- Configure main.py to include middleware and integrate the APIRouter for clean structure
- Add schemas.py to define Pydantic models for consistent data validation
- Resolve SessionMiddleware issue by correctly applying it at the FastAPI app level

issue: #2
